Try to fix issue of width between full/half screen
I have to go to sleep now, but I tried applying a change to the css based on
a cursory glance at how other websites (like Odin Project) are structured.

I want the website to reduce the amount of padding on the left and right based
on how minimized the browser is, but it just won't do it.